clean up lstm_chime_mask example code
