--- examples/model_currennt/config.cfg
@@ -1,17 +1,0 @@
-max_epochs_no_best   = 20
-max_epochs           = 100
-learning_rate        = 1e-5
-network              = network.jsn 
-train                = true
-train_file           = ../train_1_speaker.nc
-val_file             = ../val_1_speaker.nc
-weights_dist         = normal
-weights_normal_sigma = 0.1
-weights_normal_mean  = 0
-stochastic           = true
-validate_every       = 1
-parallel_sequences   = 50
-input_noise_sigma    = 0.6
-shuffle_fractions    = true
-shuffle_sequences    = false
-
