--- examples/model_currennt/run.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-/data/erdogan/prog/currennt-code/build/currennt "$@" --options_file config.cfg 
-
-
